Add files via upload
Abhishek jain <7410819778abhi@gmail.com>
1:42 AM (0 minutes ago)
to me

Sam and Alex are playing a new game where there are a number of piles, each with any number of stones in it. Players take turns removing stones from any one pile.

The number removed has to be either:

1. an integer multiple of a given number, k, where k>0.

2. if there are fewer than k stones in a pile, any number can be removed.

Determine who wins the game. Sam always starts, they both play optimally, and the last player to remove a stone wins. If Sam wins, return "Sam wins the game of n piles).", where n is the number of piles in the input. If Alex wins, return " Alex wins the game of n piles)."